+++ packages/body/test/body.test.tsx
@@ -2,21 +2,21 @@ import { render } from '@jsx-email/render';
 
 import { Body } from '../src';
 
-describe('render', () => {
+describe('render', async () => {
   beforeEach(() => {
     vi.restoreAllMocks();
     vi.resetModules();
   });
 
-  it('renders children correctly', () => {
+  it('renders children correctly', async () => {
     const testMessage = 'Test message';
-    const html = render(<Body>{testMessage}</Body>);
+    const html = await render(<Body>{testMessage}</Body>);
     expect(html).toContain(testMessage);
   });
 
-  it('passes style and other props correctly', () => {
+  it('passes style and other props correctly', async () => {
     const style = { backgroundColor: 'red' };
-    const html = render(
+    const html = await render(
       <Body style={style} data-testid="body-test">
         Test
       </Body>
@@ -25,8 +25,8 @@ describe('render', () => {
     expect(html).toContain('data-testid="body-test"');
   });
 
-  it('renders the <Body> component', () => {
-    const actualOutput = render(<Body>Lorem ipsum</Body>);
+  it('renders the <Body> component', async () => {
+    const actualOutput = await render(<Body>Lorem ipsum</Body>);
     expect(actualOutput).toMatchSnapshot();
   });
 });