+++ packages/markdown/test/markdown.test.tsx
@@ -4,9 +4,9 @@ import { render } from '@jsx-email/render';
 
 import { Markdown } from '../src';
 
-describe('<Markdown> component renders correctly', () => {
-  it('renders the markdown in the correct format for browsers', () => {
-    const actualOutput = render(
+describe('<Markdown> component renders correctly', async () => {
+  it('renders the markdown in the correct format for browsers', async () => {
+    const actualOutput = await render(
       <Markdown>
         {`# Markdown Test Document
 
@@ -61,8 +61,8 @@ console.log(\`Hello, \$\{name\}!\`);
     expect(actualOutput).toMatchSnapshot();
   });
 
-  it('renders the headers in the correct format for browsers', () => {
-    const actualOutput = render(
+  it('renders the headers in the correct format for browsers', async () => {
+    const actualOutput = await render(
       <Markdown>
         {`
 # Heading 1!
@@ -77,22 +77,22 @@ console.log(\`Hello, \$\{name\}!\`);
     expect(actualOutput).toMatchSnapshot();
   });
 
-  it('renders text in the correct format for browsers', () => {
-    const actualOutput = render(
+  it('renders text in the correct format for browsers', async () => {
+    const actualOutput = await render(
       <Markdown>{`**This is sample bold text in markdown** and *this is italic text*`}</Markdown>
     );
     expect(actualOutput).toMatchSnapshot();
   });
 
-  it('renders links in the correct format for browsers', () => {
-    const actualOutput = render(
+  it('renders links in the correct format for browsers', async () => {
+    const actualOutput = await render(
       <Markdown>{`Link to [React-email](https://react.email)`}</Markdown>
     );
     expect(actualOutput).toMatchSnapshot();
   });
 
-  it('renders lists in the correct format for browsers', () => {
-    const actualOutput = render(
+  it('renders lists in the correct format for browsers', async () => {
+    const actualOutput = await render(
       <Markdown>
         {`
 # Below is a list