+++ packages/render/test/jsx-to-string.test.tsx
@@ -17,52 +17,52 @@ const Template: FC<{
   </>
 );
 
-describe('jsx-to-string', () => {
-  it('converts a React component into HTML', () => {
-    const actualOutput = jsxToString(<Template firstName="Jim" />);
+describe('jsx-to-string', async () => {
+  it('converts a React component into HTML', async () => {
+    const actualOutput = await jsxToString(<Template firstName="Jim" />);
     expect(actualOutput).toMatchSnapshot();
   });
 
-  describe('elements', () => {
-    it('renders simple elements', () => {
-      const result = jsxToString(<div></div>);
+  describe('elements', async () => {
+    it('renders simple elements', async () => {
+      const result = await jsxToString(<div></div>);
       expect(result).toBe('<div></div>');
     });
 
-    it('renders void elements', () => {
-      const result = jsxToString(<br />);
+    it('renders void elements', async () => {
+      const result = await jsxToString(<br />);
       expect(result).toBe('<br/>');
     });
 
-    it('does not render fragments', () => {
-      const result = jsxToString(<></>);
+    it('does not render fragments', async () => {
+      const result = await jsxToString(<></>);
       expect(result).toBe('');
     });
 
-    it('invokes functional components', () => {
+    it('invokes functional components', async () => {
       const Component = (props: { text: string }) => <span className="c">{props.text}</span>;
-      const result = jsxToString(<Component text="Hello World" />);
+      const result = await jsxToString(<Component text="Hello World" />);
       expect(result).toBe('<span class="c">Hello World</span>');
     });
 
-    it('does not render null', () => {
-      expect(jsxToString(null)).toBe('');
+    it('does not render null', async () => {
+      expect(await jsxToString(null)).toBe('');
     });
 
-    it('does not render booleans', () => {
-      expect(jsxToString(true)).toBe('');
-      expect(jsxToString(false)).toBe('');
+    it('does not render booleans', async () => {
+      expect(await jsxToString(true)).toBe('');
+      expect(await jsxToString(false)).toBe('');
     });
   });
 
-  describe('properties', () => {
-    it('renders HTML attribute names and values', () => {
-      const result = jsxToString(<form action="/submit" method="post" />);
+  describe('properties', async () => {
+    it('renders HTML attribute names and values', async () => {
+      const result = await jsxToString(<form action="/submit" method="post" />);
       expect(result).toBe('<form action="/submit" method="post"></form>');
     });
 
-    it('does not render null or undefined HTML attributes', () => {
-      const result = jsxToString(
+    it('does not render null or undefined HTML attributes', async () => {
+      const result = await jsxToString(
         // @ts-expect-error because null is not a valid value for the `itemType` attribute
         // eslint-disable-next-line no-undefined
         <div itemType={null} itemProp={undefined} className={undefined} />
@@ -70,83 +70,87 @@ describe('jsx-to-string', () => {
       expect(result).toBe('<div></div>');
     });
 
-    it('ignores framework specific properties', () => {
-      const result = jsxToString(<div key="item" children={[]} />);
+    it('ignores framework specific properties', async () => {
+      const result = await jsxToString(<div key="item" children={[]} />);
       expect(result).toBe('<div></div>');
     });
 
-    it('aliases conflicting JS keywords', () => {
-      const result = jsxToString(<label className="label" htmlFor="id" />);
+    it('aliases conflicting JS keywords', async () => {
+      const result = await jsxToString(<label className="label" htmlFor="id" />);
       expect(result).toBe('<label class="label" for="id"></label>');
     });
 
-    it('lowercases camelCase attribute names', () => {
-      const result = jsxToString(<input tabIndex={-1} defaultValue="Hello" />);
+    it('lowercases camelCase attribute names', async () => {
+      const result = await jsxToString(<input tabIndex={-1} defaultValue="Hello" />);
       expect(result).toBe('<input tabindex="-1" value="Hello"/>');
     });
 
-    it('escapes HTML attribute values', () => {
-      const result = jsxToString(<img alt='"Mac & Cheese"' />);
+    it('escapes HTML attribute values', async () => {
+      const result = await jsxToString(<img alt='"Mac & Cheese"' />);
       expect(result).toBe('<img alt="&quot;Mac &amp; Cheese&quot;"/>');
     });
 
-    describe('boolean attributes', () => {
-      it('does not append boolean attributes with a falsy value', () => {
+    describe('boolean attributes', async () => {
+      it('does not append boolean attributes with a falsy value', async () => {
         // @ts-expect-error because `open` is a boolean attribute.
-        const result = jsxToString(<details hidden={false} open={0} />);
+        const result = await jsxToString(<details hidden={false} open={0} />);
         expect(result).toBe('<details></details>');
       });
 
-      it('appends boolean attributes with a truthy value', () => {
+      it('appends boolean attributes with a truthy value', async () => {
         // @ts-expect-error because `open` is a boolean attribute.
-        const result = jsxToString(<details hidden={true} open={1} />);
+        const result = await jsxToString(<details hidden={true} open={1} />);
         expect(result).toBe('<details hidden open></details>');
       });
 
-      it('renders boolean values for enumerable attributes', () => {
-        const result = jsxToString(<div contentEditable={true} spellCheck={false} />);
+      it('renders boolean values for enumerable attributes', async () => {
+        const result = await jsxToString(<div contentEditable={true} spellCheck={false} />);
         expect(result).toBe('<div contenteditable="true" spellcheck="false"></div>');
       });
     });
 
-    describe('styles', () => {
-      it('stringifies style attributes', () => {
-        const result = jsxToString(<div style={{ margin: '1em 0', padding: '0.5em 1em' }} />);
+    describe('styles', async () => {
+      it('stringifies style attributes', async () => {
+        const result = await jsxToString(<div style={{ margin: '1em 0', padding: '0.5em 1em' }} />);
         expect(result).toBe('<div style="margin:1em 0;padding:0.5em 1em"></div>');
       });
 
-      it('hyphenates style properties', () => {
-        const result = jsxToString(<div style={{ marginBottom: '1em', paddingTop: '0.5em' }} />);
+      it('hyphenates style properties', async () => {
+        const result = await jsxToString(
+          <div style={{ marginBottom: '1em', paddingTop: '0.5em' }} />
+        );
         expect(result).toBe('<div style="margin-bottom:1em;padding-top:0.5em"></div>');
       });
 
-      it('hyphenates vendor prefixed properties', () => {
-        const result = jsxToString(<div style={{ MozHyphens: 'auto', msHyphens: 'auto' }} />);
+      it('hyphenates vendor prefixed properties', async () => {
+        const result = await jsxToString(<div style={{ MozHyphens: 'auto', msHyphens: 'auto' }} />);
         expect(result).toBe('<div style="-moz-hyphens:auto;-ms-hyphens:auto"></div>');
       });
 
-      it('applies pixel units to number values', () => {
-        const result = jsxToString(<div style={{ margin: 20, padding: 10 }} />);
+      it('applies pixel units to number values', async () => {
+        const result = await jsxToString(<div style={{ margin: 20, padding: 10 }} />);
         expect(result).toBe('<div style="margin:20px;padding:10px"></div>');
       });
 
-      it('does not apply pixel values to unitless properties', () => {
-        const result = jsxToString(<div style={{ flexShrink: 1, order: 2 }} />);
+      it('does not apply pixel values to unitless properties', async () => {
+        const result = await jsxToString(<div style={{ flexShrink: 1, order: 2 }} />);
         expect(result).toBe('<div style="flex-shrink:1;order:2"></div>');
       });
 
-      it('ignores null or undefined properties', () => {
-        // @ts-expect-error because `width`, height` and `border` are not valid style properties.
-        // eslint-disable-next-line no-undefined
-        const result = jsxToString(<div style={{ border: 0, height: undefined, width: null }} />);
+      it('ignores null or undefined properties', async () => {
+        const result = await jsxToString(
+          // @ts-expect-error because `width`, height` and `border` are not valid style properties.
+          // eslint-disable-next-line no-undefined
+          <div style={{ border: 0, height: undefined, width: null }} />
+        );
         expect(result).toBe('<div style="border:0"></div>');
       });
     });
   });
 
-  describe('children', () => {
-    it('ignores empty children', () => {
-      const result = jsxToString(
+  describe('children', async () => {
+    it('ignores empty children', async () => {
+      const result = await jsxToString(
         <div>
           {/* eslint-disable-next-line no-undefined */}
           {undefined}
@@ -156,36 +160,36 @@ describe('jsx-to-string', () => {
       expect(result).toBe('<div></div>');
     });
 
-    it('renders text children', () => {
-      const result = jsxToString(<div>Hello World</div>);
+    it('renders text children', async () => {
+      const result = await jsxToString(<div>Hello World</div>);
       expect(result).toBe('<div>Hello World</div>');
     });
 
-    it('renders numeric children', () => {
-      const result = jsxToString(<div>{123}</div>);
+    it('renders numeric children', async () => {
+      const result = await jsxToString(<div>{123}</div>);
       expect(result).toBe('<div>123</div>');
     });
 
-    it('renders numeric children even if they are zero', () => {
-      const result = jsxToString(<div>{0}</div>);
+    it('renders numeric children even if they are zero', async () => {
+      const result = await jsxToString(<div>{0}</div>);
       expect(result).toBe('<div>0</div>');
     });
 
-    it('does not render boolean children', () => {
-      const a = jsxToString(<div>{true}</div>);
+    it('does not render boolean children', async () => {
+      const a = await jsxToString(<div>{true}</div>);
       expect(a).toBe('<div></div>');
 
-      const b = jsxToString(<div>{false}</div>);
+      const b = await jsxToString(<div>{false}</div>);
       expect(b).toBe('<div></div>');
     });
 
-    it('escapes text children', () => {
-      const result = jsxToString(<div>{'"Mac & Cheese"'}</div>);
+    it('escapes text children', async () => {
+      const result = await jsxToString(<div>{'"Mac & Cheese"'}</div>);
       expect(result).toBe('<div>&quot;Mac &amp; Cheese&quot;</div>');
     });
 
-    it('renders multiple children', () => {
-      const result = jsxToString(
+    it('renders multiple children', async () => {
+      const result = await jsxToString(
         <div>
           <i>Hello</i> <i>World</i>!
         </div>
@@ -193,16 +197,16 @@ describe('jsx-to-string', () => {
       expect(result).toBe('<div><i>Hello</i> <i>World</i>!</div>');
     });
 
-    it('renders nested children', () => {
-      const result = jsxToString(
+    it('renders nested children', async () => {
+      const result = await jsxToString(
         <div>{[<i key={1}>Hello</i>, [' ', [<i key={2}>World</i>, '!']]]}</div>
       );
       expect(result).toBe('<div><i>Hello</i> <i>World</i>!</div>');
     });
 
-    it('passes children to compositional components', () => {
+    it('passes children to compositional components', async () => {
       const Parent = (props: { children: ReactNode }) => <ul>{props.children}</ul>;
-      const result = jsxToString(
+      const result = await jsxToString(
         <Parent>
           <li>one</li>
           <li>two</li>
@@ -212,18 +216,18 @@ describe('jsx-to-string', () => {
       expect(result).toBe('<ul><li>one</li><li>two</li></ul>');
     });
 
-    it('it allows children provided as props', () => {
+    it('it allows children provided as props', async () => {
       const Parent = (props: { children: ReactNode }) => <ul>{props.children}</ul>;
-      const result = jsxToString(
+      const result = await jsxToString(
         <Parent children={[<li key={1}>one</li>, <li key={2}>two</li>]} />
       );
 
       expect(result).toBe('<ul><li>one</li><li>two</li></ul>');
     });
 
-    it('favours children provided as arguments over props', () => {
+    it('favours children provided as arguments over props', async () => {
       const Parent = (props: { children: ReactNode }) => <ul>{props.children}</ul>;
-      const result = jsxToString(
+      const result = await jsxToString(
         // @ts-expect-error because `children` is specified twice
         <Parent children={[<li>one</li>, <li>two</li>]}>
           <li>three</li>
@@ -233,13 +237,15 @@ describe('jsx-to-string', () => {
       expect(result).toBe('<ul><li>three</li></ul>');
     });
 
-    it('supports setting inner HTML', () => {
-      const result = jsxToString(<div dangerouslySetInnerHTML={{ __html: '<i>Hello</i>' }} />);
+    it('supports setting inner HTML', async () => {
+      const result = await jsxToString(
+        <div dangerouslySetInnerHTML={{ __html: '<i>Hello</i>' }} />
+      );
       expect(result).toBe('<div><i>Hello</i></div>');
     });
 
-    it('renders the children of fragments', () => {
-      const result = jsxToString(
+    it('renders the children of fragments', async () => {
+      const result = await jsxToString(
         <dl>
           <>
             <dt>Title</dt>