+++ packages/text/test/text.test.tsx
@@ -2,21 +2,21 @@ import { render } from '@jsx-email/render';
 
 import { Text } from '../src';
 
-describe('<Text> component', () => {
+describe('<Text> component', async () => {
   beforeEach(() => {
     vi.restoreAllMocks();
     vi.resetModules();
   });
 
-  it('renders children correctly', () => {
+  it('renders children correctly', async () => {
     const testMessage = 'Test message';
-    const html = render(<Text>{testMessage}</Text>);
+    const html = await render(<Text>{testMessage}</Text>);
     expect(html).toContain(testMessage);
   });
 
-  it('passes style and other props correctly', () => {
+  it('passes style and other props correctly', async () => {
     const style = { fontSize: '16px' };
-    const html = render(
+    const html = await render(
       <Text style={style} data-testid="text-test">
         Test
       </Text>
@@ -25,8 +25,8 @@ describe('<Text> component', () => {
     expect(html).toContain('data-testid="text-test"');
   });
 
-  it('renders correctly', () => {
-    const actualOutput = render(<Text>Lorem ipsum</Text>);
+  it('renders correctly', async () => {
+    const actualOutput = await render(<Text>Lorem ipsum</Text>);
     expect(actualOutput).toMatchSnapshot();
   });
 });